+++ app/helpers/posts_helper.rb
@@ -10,7 +10,18 @@ module PostsHelper
       I18n.t "posts.show.reshare_by", :author => post.author_name
     else
       if post.text.present?
-        truncate(post.text(:plain_text => true), :length => opts.fetch(:length, 20))
+        if opts.has_key?(:length)
+           truncate(post.text(:plain_text => true), :length => opts.fetch(:length))
+        elsif /\A(?:  # Regexp to match a Markdown header present on first line :
+              (?<setext_content>.{1,200}\n(?:={1,200}|-{1,200}))(?:\r?\n|$)           # Setext-style header
+                  |                                                                   # or
+              (?<atx_content>\#{1,6}\s.{1,200})(?:\r?\n|$)                            # Atx-style header
+               )/x   =~  post.text(:plain_text => true)
+          return setext_content unless setext_content.nil?
+          return atx_content unless atx_content.nil?
+        else
+          truncate(post.text(:plain_text => true), :length => 20 )
+        end
       elsif post.respond_to?(:photos) && post.photos.present?
         I18n.t "posts.show.photos_by", :count => post.photos.size, :author => post.author_name
       end
@@ -19,7 +30,7 @@ module PostsHelper
 
   def post_iframe_url(post_id, opts={})
     opts[:width] ||= 516
-    opts[:height] ||= 315 
+    opts[:height] ||= 315
     host = AppConfig.pod_uri.authority
    "<iframe src='#{Rails.application.routes.url_helpers.post_url(post_id, :host => host)}' width='#{opts[:width]}px' height='#{opts[:height]}px' frameBorder='0'></iframe>".html_safe
   end