+++ Changelog.md
@@ -21,6 +21,7 @@
 * Show the user if a contact is sharing with them when viewing their profile page [#2948](https://github.com/diaspora/diaspora/issues/2948)
 * Made Unicorn timeout configurable and increased the default to 90 seconds
 * Follow DiasporaHQ upon account creation is now configurable to another account [#4278](https://github.com/diaspora/diaspora/pull/4278)
+* Use first header as title in the single post view, when possible [#4256](https://github.com/diaspora/diaspora/pull/4256)
 
 # 0.1.1.0